+++ R/plot.nlmixr2AugPred.R
@@ -79,7 +79,7 @@ plot.nlmixr2AugPred <- function(x, y, ...) {
     # Skip endpoint levels without any rows (#44)
     for (.tmp in levels(droplevels(as.factor(x$Endpoint)))) {
       utils::assignInMyNamespace(".augPredEndpoint", .tmp)
-      .x <- x[x$Endpoint == .tmp, names(x) != "Endpoint"]
+      .x <- x[which(x$Endpoint == .tmp), names(x) != "Endpoint"]
       .r <- plot.nlmixr2AugPred(.x)
       for (.k in seq_along(.r)) {
         .ret[[length(.ret) + 1L]] <- .r[[.k]]